motor_voice_control/requirements.txt
openai-whisper
numpy
pyserial
sounddevice
vosk
